+++ include/sealib/graph/node.h
@@ -14,7 +14,7 @@ class Node {
  public:
     Node() : adj(0) {}
     explicit Node(std::vector<T> const &p) : adj(p) {}
-    uint getDegree() const { return adj.size(); }
+    uint getDegree() const { return static_cast<uint>(adj.size()); }
     std::vector<T> const &getAdj() const { return adj; }
     std::vector<T> &getAdj() { return adj; }
     void addAdjacency(T const &v) { adj.emplace_back(v); }
